GHCJS.Prim

Documentation

data JSVal

Constructors

JSVal ByteArray#

data JSException

Constructors

JSException JSVal String

data WouldBlockException

If a synchronous thread tries to do something that can only be done asynchronously, and the thread is set up to not continue asynchronously, it receives this exception.

Constructors

WouldBlockException

mkJSException :: JSVal -> IO JSException

fromJSString :: JSVal -> String

Low-level conversion utilities for packages that cannot depend on ghcjs-base

returns an empty string if the JSVal does not contain a string

toJSString :: String -> JSVal

toJSArray :: [JSVal] -> IO JSVal

fromJSArray :: JSVal -> IO [JSVal]

fromJSInt :: JSVal -> Int

returns zero if the JSVal does not contain a number

toJSInt :: Int -> JSVal

isNull :: JSVal -> Bool

isUndefined :: JSVal -> Bool

jsNull :: JSVal

getProp :: JSVal -> String -> IO JSVal

getProp' :: JSVal -> JSVal -> IO JSVal